Fix prepare/decide phase
With the approach we had before, we were using the same msg id for different
messages on the same connection. This works fine if all partitions are in
different connections. When more than one partition live on the same connection,
it might be the case that we overwrite a callback in pipesock, making the whole
process faulty.