Remove timeout error on vote counting
The previous timeout handling was faulty, as the vote accumulator
expects a list, not a single error tuple, so the calling process would
crash with a function clause mismatch. Remove the timeout completeley.

Some partitions might miss the 5s deadline. Even if it is a long
timeout, it's better to retry than to miss it. If simply delayed,
subsequent calls to the coordinator might receive an old message and get
confused. Reads might get votes, or votes get reads.